% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhs_dict_as_ddi.R
\name{read_dhs_dict_as_ddi}
\alias{read_dhs_dict_as_ddi}
\title{read_dhs_dict_as_ddi}
\usage{
read_dhs_dict_as_ddi(sample, ...)
}
\description{
read_dhs_dict_as_ddi
}
\note{

}
\examples{
\dontrun{}
}
\author{
Matt Gunther
}
