% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_vcal_input_data.R
\name{reshape_vcal_input_data}
\alias{reshape_vcal_input_data}
\title{reshape_vcal_input_data}
\usage{
reshape_vcal_input_data(dat, sample_tracking, ...)
}
\description{
reshape_vcal_input_data
}
\note{

}
\examples{
\dontrun{}
}
\author{
Matt Gunther
}
