% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcal_reprod.R
\name{vcal_reprod}
\alias{vcal_reprod}
\title{Reproductive Events Calendar}
\usage{
vcal_reprod(dat, vcal_reprod_recodes = NULL)
}
\arguments{
\item{dat}{A data file created by \code{reshape_vcal_input_data()}}

\item{vcal_reprod_recodes}{Optional Character: the full path to a file
called vcal_reprod_recodes.csv (contains all common and sample-specific
recodes for the Reproductive Events calendar).}
}
\description{
Create all variables related to the Reproductive Events
calendar (see details). If this calendar was not included in the sample (or
if some of the required information is not available), all
variables will \emph{still be created}, but all values will be NA.
}
\details{
The following variables will be created using case logic provided
to the function \code{dplyr::case_when()}. Please note that
\code{case_when()} returns NA through implicit logic: \emph{if a "case"
exists and is not explicitly handled here, the value NA will be returned!}

\itemize{
  \item{
    \strong{reprod_event} Numeric: a recoded version of
    \code{vcal_reprod} (usually \code{vcal_1} in the IR file).
    All values are a number, and
    sample-specific codes are harmonized using vcal_reprod_recodes.csv
  }

  \item{
    \strong{caseid_cmc} concatenates \code{caseid} and \code{cmc_month}
    (separated by \code{_})
  }

  \item{
    \strong{seq} month index, where the earliest month gets 1 and the
    final month usually gets 80 (length of the original calendar string)
  }

  \item{
    \strong{birth} Logical: TRUE if \code{vcal_reprod} is recoded 100,
    FALSE if it is any other non-missing value.
  }

  \item{
    \strong{preg} Logical: TRUE if \code{vcal_reprod} is recoded 200,
    FALSE if it is any other non-missing value.
  }

  \item{
    \strong{term} Logical: TRUE if \code{vcal_reprod} is recoded 300,
    FALSE if it is any other non-missing value.
  }

  \item{
    \strong{contr} Logical: TRUE if \code{vcal_reprod} is recoded to
    a value between 1 and 90, FALSE if it is any other non-missing value.
  }

  \item{
    \strong{eventpbt} Logical: TRUE if \code{birth}, \code{preg}, or
    \code{term} is TRUE, FALSE if they are all FALSE and non-missing.
  }

  \item{
    \strong{eventwfp} Logical: TRUE if \code{eventpbt} or \code{contr}
    is TRUE, FALSE if both are FALSE and non-missing.
  }

  \item{
    \strong{switch} Logical: TRUE if \code{vcal_reprod} is recoded 1:90 and
    the previous month's value for \code{vcal_reprod} is recoded 0:90
    \emph{but}
    not the same value as the current month. FALSE in all other cases where
    \emph{both} the current and previous month are available (e.g. excludes
    the first month and any missing value) and \emph{either} falls beyond
    the required range.
  }

  \item{
    \strong{switch_new} Logical: TRUE if \code{switch} is TRUE
    \emph{and} \code{vcal_reprod} in the previous month \strong{not}
    recoded 0.
    FALSE if \code{switch} is TRUE  \emph{and} \code{vcal_reprod} in the
    previous month \strong{is} recoded 0.
  }

  \item{
    \strong{disc_event} see \code{vcal_discont()}
  }

  \item{
    \strong{total_preg} Integer: the total number of months per person
    where \code{preg} is TRUE (NA if \code{preg} is not available). All
    months for the person reflect the same total (this is not a cumulative
    sum).
  }

  \item{
    \strong{count_birth} Integer: the total number of months per person
    where \code{birth} is TRUE (NA if \code{birth} is not available). All
    months for the person reflect the same total (this is not a cumulative
    sum).
  }

  \item{
    \strong{count_term} Integer: the total number of months per person
    where \code{term} is TRUE (NA if \code{term} is not available). All
    months for the person reflect the same total (this is not a cumulative
    sum).
  }

  \item{
    \strong{contr_duration} Integer: the total number of months per person
    where \code{contr} is TRUE (NA if \code{contr} is not available). All
    months for the person reflect the same total (this is not a cumulative
    sum).
  }

  \item{
  \strong{trunc} Logical: TRUE for all months in a continuous
    pregnancy if the outcome of the pregnancy (birth or termination)
    happened
    after the date of the interview. FALSE for all months in a continuous
    pregnancy otherwise. Not available (NA) for all months that were not
    part
    of a continuous pregnancy.
  }

  \item{
    \strong{prfirst} Logical: TRUE for all months in a continuous
    pregnancy that overlaps with the first month in the recall timeline.
    FALSE
    for all months in a continuous pregnancy otherwise. Not available (NA)
    for months that were not part of a continuous pregnancy.
  }

  \item{
    \strong{preg_length} Integer: the total number of months included
    in a single continuous pregnancy where \emph{both} \code{trunc} and
    \code{prfirst} are FALSE. All months in the pregnancy reflect the
    same total (this is not a cumulative sum). Not available (NA) for
    months that were not part of a continuous pregnancy.
  }

  \item{
    \strong{preg_flag} Logical: TRUE if \code{preg_length} is longer
    than 9 months. Not available (NA) for months where \code{preg_length}
    is not available.
  }
}
}
\author{
Matt Gunther
}
