% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcal_term.R
\name{vcal_term}
\alias{vcal_term}
\title{Pregnancy Termination Calendar}
\usage{
vcal_term(dat)
}
\arguments{
\item{dat}{A data file created by \code{vcal_reprod()} (may be passed to
any other function starting with "vcal" first).}
}
\description{
Create all variables related to the Pregnancy Termination
calendar
(see details). If this calendar was not included in the sample (or if some
of the required information is not available), all variables will still be
 created, but all values will be NA.
}
\details{
The following variables will be created using case logic provided
to the function \code{dplyr::case_when()}. Please note that
\code{case_when()} returns NA through implicit logic: \emph{if a "case"
exists and is not explicitly handled here, the value NA will be returned!}
\itemize{
  \item{
    \strong{abort} Logical: TRUE if \code{vcal_term} is "A". FALSE if
    \code{vcal_term} is any other non-missing character. Not available if
    \code{vcal_term} is blank (e.g. " ") or not available (NA).
  }
  \item{
    \strong{count_abort} Integer: total number of months per person where
    \code{abort} is TRUE. Not available if \code{vcal_term} is not available
    (NA).
  }
  \item{
    \strong{miscar} Logical: TRUE if \code{vcal_term} is "M" or "C". FALSE
    if \code{vcal_term} is any other non-missing character. Not available if
    \code{vcal_term} is blank (e.g. " ") or not available (NA).
  }
  \item{
    \strong{count_miscar} Integer: total number of months per person where
    \code{miscar} is TRUE. Not available if \code{vcal_term} is not
    available (NA).
  }
  \item{
    \strong{sbirth} Logical: TRUE if \code{vcal_term} is "S". FALSE if
    \code{vcal_term} is any other non-missing character. Not available if
    \code{vcal_term} is blank (e.g. " ") or not available (NA).
  }
  \item{
    \strong{count_sbirth} Integer: total number of months per person where
    \code{sbirth} is TRUE. Not available if \code{vcal_term} is not
    available (NA).
  }
}
}
\author{
Matt Gunther
}
