% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcal_input_data.R
\name{get_vcal_input_data}
\alias{get_vcal_input_data}
\title{get_vcal_input_data}
\usage{
get_vcal_input_data(ddi, ...)
}
\description{
get_vcal_input_data
}
\note{

}
\examples{
\dontrun{}
}
\author{
Matt Gunther
}
